Move metabrowse stuff to a separate trait
Aiming at having a dummy subsitute in 2.13
