+++ modules/scala/scala-interpreter/src/main/scala/almond/ScalaInterpreter.scala
@@ -1,20 +1,16 @@
 package almond
 
-import java.io.File
-import java.net.URI
 import java.nio.charset.StandardCharsets.UTF_8
-import java.nio.file.{Files, Path, Paths}
-import java.util.UUID
+import java.nio.file.{Files, Path}
 
 import almond.api.JupyterApi
 import almond.api.helpers.Display
-import almond.channels.ConnectionParameters
 import almond.internals._
 import almond.interpreter._
 import almond.interpreter.api.{CommHandler, DisplayData, OutputHandler}
 import almond.interpreter.input.InputManager
 import almond.interpreter.util.CancellableFuture
-import almond.logger.{Logger, LoggerContext}
+import almond.logger.LoggerContext
 import almond.protocol.KernelInfo
 import ammonite.interp.{Parsers, Preprocessor}
 import ammonite.ops.read
@@ -26,15 +22,13 @@ import coursier.almond.tmp.Tmp
 import fastparse.Parsed
 import io.github.soc.directories.ProjectDirectories
 import jupyter.{Displayer, Displayers}
-import metabrowse.server.{MetabrowseServer, Sourcepath}
 import pprint.{TPrint, TPrintColors}
 
 import scala.collection.mutable
 import scala.concurrent.{Await, ExecutionContext}
 import scala.concurrent.duration.Duration
 import scala.reflect.ClassTag
-import scala.tools.nsc.Global
-import scala.util.{Failure, Random, Success}
+import scala.util.{Failure, Success}
 
 final class ScalaInterpreter(
   updateBackgroundVariablesEcOpt: Option[ExecutionContext] = None,
@@ -49,95 +43,18 @@ final class ScalaInterpreter(
   codeWrapper: Preprocessor.CodeWrapper = Preprocessor.CodeClassWrapper,
   initialColors: Colors = Colors.Default,
   initialClassLoader: ClassLoader = Thread.currentThread().getContextClassLoader,
-  logCtx: LoggerContext = LoggerContext.nop,
-  metabrowse: Boolean = false,
-  metabrowseHost: String = "localhost",
-  metabrowsePort: Int = -1,
+  val logCtx: LoggerContext = LoggerContext.nop,
+  val metabrowse: Boolean = false,
+  val metabrowseHost: String = "localhost",
+  val metabrowsePort: Int = -1,
   lazyInit: Boolean = false,
   trapOutput: Boolean = false,
   disableCache: Boolean = false
-) extends Interpreter { scalaInterp =>
+) extends Interpreter with ScalaInterpreterInspections { scalaInterp =>
 
   private val log = logCtx(getClass)
 
-  @volatile private var metabrowseServerOpt0 = Option.empty[(MetabrowseServer, Int, String)]
-  private val metabrowseServerCreateLock = new Object
-
-  private def metabrowseServerOpt() =
-    if (metabrowse)
-      metabrowseServerOpt0.orElse {
-        metabrowseServerCreateLock.synchronized {
-          metabrowseServerOpt0.orElse {
-            metabrowseServerOpt0 = Some(createMetabrowseServer())
-            metabrowseServerOpt0
-          }
-        }
-      }
-    else
-      None
-
-  private def createMetabrowseServer() = {
-
-    if (metabrowse && !sys.props.contains("org.jboss.logging.provider") && !sys.props.get("almond.adjust.jboss.logging.provider").contains("0")) {
-      log.info("Setting Java property org.jboss.logging.provider to slf4j")
-      sys.props("org.jboss.logging.provider") = "slf4j"
-    }
-
-    val port =
-      if (metabrowsePort > 0)
-        metabrowsePort
-      else
-        ConnectionParameters.randomPort()
-
-    val server = new MetabrowseServer(
-      host = metabrowseHost,
-      port = port
-      // FIXME Pass custom logger?
-    )
-
-    val windowName = {
-      val id = math.abs(Random.nextInt().toLong)
-      s"almond-metabrowse-$id"
-    }
-
-    val baseSourcepath = ScalaInterpreter.baseSourcePath(
-      frames()
-        .last
-        .classloader
-        .getParent,
-      log
-    )
-
-    val sourcePath = {
-
-      import ScalaInterpreter.SourcepathOps
-
-      val sessionJars = frames()
-        .flatMap(_.classpath)
-        .collect {
-          // FIXME We're ignoring jars-in-jars of standalone bootstraps of coursier in particular
-          case p if p.getProtocol == "file" =>
-            Paths.get(p.toURI)
-        }
-
-      val (sources, other) = sessionJars
-        .partition(_.getFileName.toString.endsWith("-sources.jar"))
-
-      Sourcepath(other, sources) :: baseSourcepath
-    }
-
-    log.info(s"Starting metabrowse server at http://$metabrowseHost:$port")
-    log.info(
-      "Initial source path\n  Classpath\n" +
-        sourcePath.classpath.map("    " + _).mkString("\n") +
-        "\n  Sources\n" +
-        sourcePath.sources.map("    " + _).mkString("\n")
-    )
-    server.start(sourcePath)
-
-    (server, port, windowName)
-  }
-
+  def pressy = ammInterp.compilerManager.pressy.compiler
 
   private val colors0 = Ref[Colors](initialColors)
   private val history0 = new History(Vector())
@@ -197,9 +114,11 @@ final class ScalaInterpreter(
     else
       new Storage.Folder(os.Path(ProjectDirectories.from(null, null, "Almond").cacheDir) / "ammonite")
 
-  private val frames = Ref(List(Frame.createInitial(initialClassLoader)))
-  private val sess0 = new SessionApiImpl(frames)
+  private val frames0 = Ref(List(Frame.createInitial(initialClassLoader)))
+  private val sess0 = new SessionApiImpl(frames0)
   private var currentLine0 = 0
+
+  def frames(): List[Frame] = frames0()
 
   private val printer0 = Printer(
     capture.out,
@@ -392,9 +311,9 @@ final class ScalaInterpreter(
             (almond.api.JupyterAPIHolder.getClass.getName.stripSuffix("$"), "kernel", jupyterApi)
           ),
           colors = Ref(Colors.Default),
-          getFrame = () => frames().head,
+          getFrame = () => frames0().head,
           createFrame = () => {
-            val f = sess0.childFrame(frames().head); frames() = f :: frames(); f
+            val f = sess0.childFrame(frames0().head); frames0() = f :: frames0(); f
           },
           replCodeWrapper = codeWrapper,
           scriptCodeWrapper = codeWrapper,
@@ -650,7 +569,7 @@ final class ScalaInterpreter(
 
     val (newPos, completions0, _) = ammInterp.compilerManager.complete(
       pos,
-      frames().head.imports.toString(),
+      frames0().head.imports.toString(),
       code
     )
 
@@ -664,64 +583,6 @@ final class ScalaInterpreter(
       completions.map(_.trim).distinct
     )
   }
-
-  override def inspect(code: String, pos: Int, detailLevel: Int): Option[Inspection] =
-    metabrowseServerOpt().flatMap {
-      case (metabrowseServer, metabrowsePort0, metabrowseWindowId) =>
-        val pressy = ammInterp.compilerManager.pressy.compiler
-
-        val prefix = frames().head.imports.toString() + newLine + "object InspectWrapper{" + newLine
-        val suffix = newLine + "}"
-        val allCode = prefix + code + suffix
-        val index = prefix.length + pos
-
-        val currentFile = new scala.reflect.internal.util.BatchSourceFile(
-          ammonite.interp.Compiler.makeFile(allCode.getBytes, name = "Current.sc"),
-          allCode
-        )
-
-        val r = new scala.tools.nsc.interactive.Response[Unit]
-        pressy.askReload(List(currentFile), r)
-        r.get.swap match {
-          case Left(e) =>
-            log.warn(s"Error loading '${code.take(pos)}|${code.drop(pos)}' into presentation compiler", e)
-            None
-          case Right(()) =>
-            val r0 = new scala.tools.nsc.interactive.Response[pressy.Tree]
-            pressy.askTypeAt(new scala.reflect.internal.util.OffsetPosition(currentFile, index), r0)
-            r0.get.swap match {
-              case Left(e) =>
-                log.debug(s"Getting type info for '${code.take(pos)}|${code.drop(pos)}' via presentation compiler", e)
-                None
-              case Right(tree) =>
-
-                val r0 = pressy.askForResponse(() => metabrowseServer.urlForSymbol(pressy)(tree.symbol))
-                r0.get.swap match {
-                  case Left(e) =>
-                    log.warn(s"Error loading '${code.take(pos)}|${code.drop(pos)}' into presentation compiler", e)
-                    None
-                  case Right(relUrlOpt) =>
-                    log.debug(s"url of $tree: $relUrlOpt")
-                    val urlOpt = relUrlOpt.map(relUrl => s"http://$metabrowseHost:$metabrowsePort0/$relUrl")
-
-                    val typeStr = ScalaInterpreter.typeOfTree(pressy)(tree).getOrElse(tree.toString)
-
-                    import scalatags.Text.all._
-
-                    val typeHtml0 = pre(typeStr)
-                    val typeHtml: Frag = urlOpt.fold(typeHtml0) { url =>
-                      a(href := url, target := metabrowseWindowId, typeHtml0)
-                    }
-
-                    val res = Inspection.fromDisplayData(
-                      DisplayData.html(typeHtml.toString)
-                    )
-
-                    Some(res)
-                }
-            }
-        }
-      }
 
   def kernelInfo() =
     KernelInfo(
@@ -744,10 +605,7 @@ final class ScalaInterpreter(
     )
 
   override def shutdown(): Unit =
-    for ((metabrowseServer, _, _) <- metabrowseServerOpt0) {
-      log.info("Stopping metabrowse server")
-      metabrowseServer.stop()
-    }
+    inspectionsShutdown()
 
 }
 
@@ -818,79 +676,6 @@ object ScalaInterpreter {
         ex, colors.error(), fansi.Attr.Reset, colors.literal()
       ))
     )
-
-  private def baseSourcePath(loader: ClassLoader, log: Logger): Sourcepath = {
-
-    lazy val javaDirs = {
-      val l = Seq(sys.props("java.home")) ++
-        sys.props.get("java.ext.dirs").toSeq.flatMap(_.split(File.pathSeparator)).filter(_.nonEmpty) ++
-        sys.props.get("java.endorsed.dirs").toSeq.flatMap(_.split(File.pathSeparator)).filter(_.nonEmpty)
-      l.map(_.stripSuffix("/") + "/")
-    }
-
-    def isJdkJar(uri: URI): Boolean =
-      uri.getScheme == "file" && {
-        val path = new File(uri).getAbsolutePath
-        javaDirs.exists(path.startsWith)
-      }
-
-    def classpath(cl: ClassLoader): Stream[java.net.URL] = {
-      if (cl == null)
-        Stream()
-      else {
-        val cp = cl match {
-          case u: java.net.URLClassLoader => u.getURLs.toStream
-          case _ => Stream()
-        }
-
-        cp #::: classpath(cl.getParent)
-      }
-    }
-
-    val baseJars = classpath(loader)
-      .map(_.toURI)
-      // assuming the JDK on the YARN machines already have those
-      .filter(u => !isJdkJar(u))
-      .map(Paths.get)
-      .toList
-
-    log.info(
-      "Found base JARs:\n" +
-        baseJars.sortBy(_.toString).map("  " + _).mkString("\n") +
-        "\n"
-    )
-
-    val (baseSources, baseOther) = baseJars
-      .partition(_.getFileName.toString.endsWith("-sources.jar"))
-
-    Sourcepath(baseOther, baseSources)
-  }
-
-  private implicit class SourcepathOps(private val p: Sourcepath) extends AnyVal {
-    def ::(other: Sourcepath): Sourcepath =
-      Sourcepath(other.classpath ::: p.classpath, other.sources ::: p.sources)
-  }
-
-  // from https://github.com/scalameta/metals/blob/cec8b98cba23110d5b2919d9879c78d3b0146ab2/metaserver/src/main/scala/scala/meta/languageserver/providers/HoverProvider.scala#L34-L51
-  // (via https://github.com/almond-sh/almond/pull/235#discussion_r222696661)
-  private def typeOfTree(c: Global)(t: c.Tree): Option[String] = {
-    import c._
-
-    val stringOrTree = t match {
-      case t: DefDef => Right(t.symbol.asMethod.info.toLongString)
-      case t: ValDef if t.tpt != null => Left(t.tpt)
-      case t: ValDef if t.rhs != null => Left(t.rhs)
-      case x => Left(x)
-    }
-
-    stringOrTree match {
-      case Right(string) => Some(string)
-      case Left(null) => None
-      case Left(tree) if tree.tpe ne NoType => Some(tree.tpe.widen.toString)
-      case _ => None
-    }
-
-  }
 
   private class Foo
 }
